refactor: move chain to its own module
