+++ config/src/chain.rs
@@ -1,0 +1,94 @@
+use serde::{Deserialize, Deserializer, Serialize};
+use std::{fmt, str::FromStr};
+
+/// Either a named or chain id or the actual id value
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+#[serde(untagged)]
+pub enum Chain {
+    #[serde(serialize_with = "super::from_str_lowercase::serialize")]
+    Named(ethers_core::types::Chain),
+    Id(u64),
+}
+
+impl Chain {
+    /// The id of the chain
+    pub fn id(&self) -> u64 {
+        match self {
+            Chain::Named(chain) => *chain as u64,
+            Chain::Id(id) => *id,
+        }
+    }
+}
+
+impl fmt::Display for Chain {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Chain::Named(chain) => chain.fmt(f),
+            Chain::Id(id) => {
+                if let Ok(chain) = ethers_core::types::Chain::try_from(*id) {
+                    chain.fmt(f)
+                } else {
+                    id.fmt(f)
+                }
+            }
+        }
+    }
+}
+
+impl From<ethers_core::types::Chain> for Chain {
+    fn from(id: ethers_core::types::Chain) -> Self {
+        Chain::Named(id)
+    }
+}
+
+impl From<u64> for Chain {
+    fn from(id: u64) -> Self {
+        Chain::Id(id)
+    }
+}
+
+impl From<Chain> for u64 {
+    fn from(c: Chain) -> Self {
+        match c {
+            Chain::Named(c) => c as u64,
+            Chain::Id(id) => id,
+        }
+    }
+}
+
+impl FromStr for Chain {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        if let Ok(chain) = ethers_core::types::Chain::from_str(s) {
+            Ok(Chain::Named(chain))
+        } else {
+            s.parse::<u64>()
+                .map(Chain::Id)
+                .map_err(|_| format!("Expected known chain or integer, found: {}", s))
+        }
+    }
+}
+
+impl<'de> Deserialize<'de> for Chain {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        #[derive(Deserialize)]
+        #[serde(untagged)]
+        enum ChainId {
+            Named(String),
+            Id(u64),
+        }
+
+        match ChainId::deserialize(deserializer)? {
+            ChainId::Named(s) => {
+                s.to_lowercase().parse().map(Chain::Named).map_err(serde::de::Error::custom)
+            }
+            ChainId::Id(id) => Ok(ethers_core::types::Chain::try_from(id)
+                .map(Chain::Named)
+                .unwrap_or_else(|_| Chain::Id(id))),
+        }
+    }
+}
