+++ config/src/lib.rs
@@ -3,20 +3,9 @@ extern crate core;
 
 use std::{
     borrow::Cow,
-    fmt,
     path::{Path, PathBuf},
     str::FromStr,
 };
-
-use figment::{
-    providers::{Env, Format, Serialized, Toml},
-    value::{Dict, Map},
-    Error, Figment, Metadata, Profile, Provider,
-};
-// reexport so cli types can implement `figment::Provider` to easily merge compiler arguments
-pub use figment;
-use semver::Version;
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
 use crate::caching::StorageCachingConfig;
 use ethers_core::types::{Address, U256};
@@ -29,8 +18,14 @@ use ethers_solc::{
     ConfigurableArtifacts, EvmVersion, Project, ProjectPathsConfig, Solc, SolcConfig,
 };
 use eyre::{ContextCompat, WrapErr};
-use figment::{providers::Data, value::Value};
+use figment::{
+    providers::{Data, Env, Format, Serialized, Toml},
+    value::{Dict, Map, Value},
+    Error, Figment, Metadata, Profile, Provider,
+};
 use inflector::Inflector;
+use semver::Version;
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
 // Macros useful for creating a figment.
 mod macros;
@@ -40,6 +35,11 @@ pub mod utils;
 pub use crate::utils::*;
 
 pub mod caching;
+mod chain;
+pub use chain::Chain;
+
+// reexport so cli types can implement `figment::Provider` to easily merge compiler arguments
+pub use figment;
 
 /// Foundry configuration
 ///
@@ -1287,85 +1287,7 @@ impl BasicConfig {
     }
 }
 
-/// Either a named or chain id or the actual id value
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[serde(untagged)]
-pub enum Chain {
-    #[serde(serialize_with = "from_str_lowercase::serialize")]
-    Named(ethers_core::types::Chain),
-    Id(u64),
-}
-
-impl Chain {
-    /// The id of the chain
-    pub fn id(&self) -> u64 {
-        match self {
-            Chain::Named(chain) => *chain as u64,
-            Chain::Id(id) => *id,
-        }
-    }
-}
-
-impl fmt::Display for Chain {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Chain::Named(chain) => chain.fmt(f),
-            Chain::Id(id) => {
-                if let Ok(chain) = ethers_core::types::Chain::try_from(*id) {
-                    chain.fmt(f)
-                } else {
-                    id.fmt(f)
-                }
-            }
-        }
-    }
-}
-
-impl From<ethers_core::types::Chain> for Chain {
-    fn from(id: ethers_core::types::Chain) -> Self {
-        Chain::Named(id)
-    }
-}
-
-impl From<u64> for Chain {
-    fn from(id: u64) -> Self {
-        Chain::Id(id)
-    }
-}
-
-impl From<Chain> for u64 {
-    fn from(c: Chain) -> Self {
-        match c {
-            Chain::Named(c) => c as u64,
-            Chain::Id(id) => id,
-        }
-    }
-}
-
-impl<'de> Deserialize<'de> for Chain {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        #[derive(Deserialize)]
-        #[serde(untagged)]
-        enum ChainId {
-            Named(String),
-            Id(u64),
-        }
-
-        match ChainId::deserialize(deserializer)? {
-            ChainId::Named(s) => {
-                s.to_lowercase().parse().map(Chain::Named).map_err(serde::de::Error::custom)
-            }
-            ChainId::Id(id) => Ok(ethers_core::types::Chain::try_from(id)
-                .map(Chain::Named)
-                .unwrap_or_else(|_| Chain::Id(id))),
-        }
-    }
-}
-
-mod from_str_lowercase {
+pub(crate) mod from_str_lowercase {
     use std::str::FromStr;
 
     use serde::{Deserialize, Deserializer, Serializer};
